Update the config to reflect the refactoring done to the application.
Adds a [zone] section for configuring the alarm(s) - in a multi-node
setup, it would be common to have multi-nodes emitting information. This
should help set guard_hours for a zone + identify which zone is emitting
the information

Adds [dropbox] and [pushbullet] sections respectively

Removes the MAC section, and wipes several lines of commentary from the
config. Explainers are in the README -- for now. There's more
documentation work to be had here. I did go a bit delete happy.